pyproject.toml
[tool.poetry]
name = "content_fetcher"
version = "0.1.0"
description = "A Python app to fetch watchlists from various services"
authors = ["erix <erix@icloud.com>"]
packages = [{include = "content_fetcher", from = "src"}]


[tool.poetry.dependencies]
python = "^3.9"
plexapi = "^4.13.2"
trakt-py = "^4.4.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
mypy = "^1.3.0"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
